server/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Flask==1.1.2
gunicorn==19.10.0
pandas==0.24.2
opencv-python==4.2.0.32
flask-cors==3.0.8
flask-socketio==4.3.1
python-engineio==3.13.1